requirements.txt
Django==1.7
Pillow==2.5.3
dj-database-url==0.3.0
dj-static==0.0.6
django-allauth==0.18.0
django-appconf==0.6
django-compressor==1.4
django-countries==2.1.2
django-crispy-forms==1.4.0
django-toolbelt==0.0.1
gunicorn==19.1.1
nodeenv==0.11.1
oauthlib==0.6.3
psycopg2==2.5.3
python3-openid==3.0.4
requests==2.4.1
requests-oauthlib==0.4.1
selenium==2.42.1
six==1.7.3
static3==0.5.1
